🚚 api: second route: move session finalization route
 from lib to src
