+++ api/lib/application/sessions/index.js
@@ -151,41 +151,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'PUT',
-      path: '/api/sessions/{id}/finalization',
-      config: {
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          params: Joi.object({
-            id: identifiersType.sessionId,
-          }),
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'examiner-global-comment': Joi.string().optional().allow(null).allow('').max(500),
-                'has-incident': Joi.boolean().required(),
-                'has-joining-issue': Joi.boolean().required(),
-              },
-            },
-          }),
-        },
-        pre: [
-          {
-            method: authorization.verifySessionAuthorization,
-            assign: 'authorizationCheck',
-          },
-        ],
-        handler: sessionController.finalize,
-        tags: ['api', 'sessions'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle permet de finaliser une session de certification afin de la signaler comme terminée',
-        ],
-      },
-    },
-    {
       method: 'POST',
       path: '/api/sessions/{id}/certification-candidates/import',
       config: {
